pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "HubSpot CRM Agent — Azure OpenAI + MCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.30.0",
    "mcp[cli]>=1.8.0",
    "prompt-toolkit>=3.0.51",
    "python-dotenv>=1.1.0",
    "httpx>=0.27.0",
]
